pyproject.toml
[project]
name = "stock-advise"
version = "0.1.0"
description = "Stock intelligence and alert system powered by Claude AI"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "feedparser>=6.0",
    "requests>=2.31",
    "slack-sdk>=3.30",
    "apscheduler>=3.10,<4",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
stock-advise = "scripts.run:main"
